Fix bug in Tensor.set(...).

+++ src/volgyerdo.math/classes/volgyerdo/math/tensor/FloatTensor.java
@@ -82,7 +82,7 @@ class FloatTensor extends Tensor {
         checkClass(tensor);
         checkDimensionCount(tensor.dimensions);
         checkDimensions(tensor);
-        System.arraycopy(((ByteTensor) tensor).values, 0, values, 0, values.length);
+        System.arraycopy(((FloatTensor) tensor).values, 0, values, 0, values.length);
     }
 
     @Override
